pyproject.toml
[project]
name = "credit-risk-monitoring-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-providers-postgres>=5.14.0",
    "apache-airflow-providers-snowflake>=3.3.0",
    "apache-airflow[amazon,apache-spark]==2.8.1",
    "boto3>=1.42.69",
    "flask-session<0.6.0",
    "fredapi>=0.5.2",
    "pandas>=2.0.0,<3.0.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=10.0.1",
    "pyspark>=4.1.1",
    "requests>=2.32.5",
    "s3fs>=2026.2.0",
    "snowflake-connector-python[pandas]>=3.12.0",
    "snowflake-sqlalchemy>=1.6.0",
    "wtforms<3.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "notebook>=7.5.5",
]
